+++ core/src/utils/chart-svgelements.e2e.ts
@@ -284,6 +284,5 @@ describe('svgelements', () => {
         expect(textEl.getAttribute('y')).toEqual("72.5");
     });
     
-  
 });